+++ backend/src/main/java/wooteco/retrospective/application/pair/PairService.java
@@ -1,15 +1,15 @@
-package wooteco.retrospective.application;
+package wooteco.retrospective.application.pair;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import wooteco.retrospective.application.dto.PairResponseDto;
 import wooteco.retrospective.dao.attendance.AttendanceDao;
-import wooteco.retrospective.dao.attendance.TimeDao;
 import wooteco.retrospective.domain.attendance.Attendance;
 import wooteco.retrospective.domain.attendance.Time;
 import wooteco.retrospective.domain.dao.PairDao;
 import wooteco.retrospective.domain.pair.Pairs;
 import wooteco.retrospective.domain.pair.member.ShuffledAttendances;
+import wooteco.retrospective.infrastructure.dao.attendance.TimeDao;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
