+++ backend/src/test/java/wooteco/retrospective/application/PairServiceTest.java
@@ -1,6 +1,7 @@
 package wooteco.retrospective.application;
 
 import org.junit.jupiter.api.Test;
+import wooteco.retrospective.application.pair.PairService;
 
 
 class PairServiceTest {
